ingestor/pyproject.toml
[project]
name = "algotrade"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.133.1",
    "pytest>=9.0.2",
    "python-dateutil>=2.9.0.post0",
    "uvicorn>=0.41.0",
]
